requirements.txt
abnf==2.2.0
ag2==0.9.7
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aiohttp-cors==0.8.1
aiosignal==1.4.0
alembic==1.16.5
altair==5.5.0
annotated-types==0.7.0
anthropic==0.71.0
anyio==4.11.0
appnope==0.1.4
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
arxiv==2.2.0
asttokens==3.0.1
async-lru==2.0.5
async-timeout==4.0.3
asyncer==0.0.8
asyncio==4.0.0
attrs==25.4.0
Authlib==1.6.5
autogen==0.9.7
autogenstudio==0.1.5
azure-common==1.1.28
azure-core==1.36.0
azure-keyvault-secrets==4.10.0
azure-mgmt-compute==37.0.1
azure-mgmt-core==1.6.0
azure-mgmt-resource==24.0.0
azure-mgmt-storage==24.0.0
azure-storage-blob==12.27.1
babel==2.17.0
backoff==2.2.1
backports.asyncio.runner==1.2.0
bcrypt==5.0.0
beautifulsoup4==4.14.3
bleach==6.2.0
blinker==1.9.0
brotli==1.2.0
build==1.4.0
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.3
chromadb==1.4.1
cint==1.0.0
click==8.1.8
cloudpickle==3.1.2
cmake==4.1.2
coloredlogs==15.0.1
colorful==0.5.7
comm==0.2.3
cryptography==46.0.3
cssselect2==0.8.0
dataclasses-json==0.6.7
debugpy==1.8.19
decorator==5.2.1
deepeval==3.7.2
defusedxml==0.7.1
Deprecated==1.3.1
diskcache==5.6.3
distlib==0.4.0
distro==1.9.0
docker==7.1.0
docstring_parser==0.17.0
durationpy==0.10
eval_type_backport==0.2.2
exceptiongroup==1.3.0
execnet==2.1.2
executing==2.2.1
fastapi==0.119.1
fastjsonschema==2.21.2
feedparser==6.0.12
fickling==0.1.4
filelock==3.19.1
flatbuffers==25.9.23
fonttools==4.60.2
fqdn==1.5.1
frozenlist==1.8.0
fsspec==2025.9.0
gitdb==4.0.12
GitPython==3.1.45
google-adk==1.18.0
google-api-core==2.27.0
google-api-python-client==2.187.0
google-auth==2.41.1
google-auth-httplib2==0.2.1
google-cloud-aiplatform==1.126.1
google-cloud-appengine-logging==1.7.0
google-cloud-audit-log==0.4.0
google-cloud-bigquery==3.38.0
google-cloud-bigtable==2.34.0
google-cloud-core==2.4.3
google-cloud-discoveryengine==0.13.12
google-cloud-logging==3.12.1
google-cloud-monitoring==2.28.0
google-cloud-resource-manager==1.15.0
google-cloud-secret-manager==2.25.0
google-cloud-spanner==3.59.0
google-cloud-speech==2.34.0
google-cloud-storage==3.5.0
google-cloud-trace==1.17.0
google-crc32c==1.7.1
google-genai==1.46.0
google-resumable-media==2.7.2
googleapis-common-protos==1.71.0
gql==4.0.0
graphql-core==3.2.7
graphviz==0.21
greenlet==3.2.4
grpc-google-iam-v1==0.14.3
grpc-interceptor==0.15.4
grpcio==1.76.0
grpcio-status==1.76.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httplib2==0.31.0
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==1.3.4
humanfriendly==10.0
idna==3.11
importlib_metadata==8.7.0
importlib_resources==6.5.2
iniconfig==2.1.0
intervaltree==3.1.0
invoke==2.2.1
ipykernel==6.31.0
ipython==8.18.1
isodate==0.7.2
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.11.1
json5==0.13.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.5.1
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
kaitaistruct==0.11
kubernetes==35.0.0
langchain==0.3.27
langchain-community==0.3.31
langchain-core==0.3.83
langchain-openai==0.3.35
langchain-text-splitters==0.3.11
langgraph==0.6.11
langgraph-checkpoint==2.1.2
langgraph-prebuilt==0.6.5
langgraph-sdk==0.2.9
langsmith==0.4.37
lark==1.3.1
lightning==2.5.5
lightning-utilities==0.15.2
loguru==0.7.3
Mako==1.3.10
Markdown==3.9
markdown-it-py==3.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
matplotlib-inline==0.2.1
mdurl==0.1.2
mistralai==1.9.11
mistune==3.2.0
mmh3==5.2.0
mpmath==1.3.0
msgpack==1.1.2
msrest==0.7.1
multidict==6.7.0
mypy_extensions==1.1.0
narwhals==2.9.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.2.1
ninja==1.13.0
notebook_shim==0.2.4
numpy==1.26.4
oauthlib==3.3.1
ollama==0.6.1
onnxruntime==1.19.2
openai==2.6.0
opencensus==0.11.4
opencensus-context==0.1.3
opentelemetry-api==1.38.0
opentelemetry-exporter-gcp-logging==1.11.0a0
opentelemetry-exporter-gcp-monitoring==1.11.0a0
opentelemetry-exporter-gcp-trace==1.11.0
opentelemetry-exporter-otlp-proto-common==1.38.0
opentelemetry-exporter-otlp-proto-grpc==1.38.0
opentelemetry-exporter-otlp-proto-http==1.37.0
opentelemetry-exporter-prometheus==0.59b0
opentelemetry-proto==1.38.0
opentelemetry-resourcedetector-gcp==1.11.0a0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
orjson==3.11.5
ormsgpack==1.11.0
overrides==7.7.0
packaging==25.0
pandas==2.3.3
pandoc==2.4
pandocfilters==1.5.1
parso==0.8.5
pdflatex==0.1.3
pdfminer.six==20250506
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.4.0
playwright==1.55.0
pluggy==1.6.0
plumbum==1.10.0
ply==3.11
polyfile-weave==0.5.7
portalocker==3.2.0
posthog==5.4.0
prometheus_client==0.23.1
prompt_toolkit==3.0.52
propcache==0.4.1
proto-plus==1.26.1
protobuf==6.33.0
psutil==7.2.1
psycopg==3.2.11
ptyprocess==0.7.0
pure_eval==0.2.3
py-spy==0.4.1
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyautogen==0.9.0
pybase64==1.4.3
pycparser==2.23
pydantic==2.12.3
pydantic-settings==2.11.0
pydantic_core==2.41.4
pydeck==0.9.1
pydyf==0.11.0
pyee==13.0.0
pyfiglet==1.0.4
Pygments==2.19.2
pypandoc==1.16.2
pyparsing==3.2.5
pyphen==0.17.2
PyPika==0.50.0
pyproject_hooks==1.2.0
pyquil==4.17.0
pytest==8.4.2
pytest-asyncio==1.2.0
pytest-repeat==0.9.4
pytest-rerunfailures==16.0.1
pytest-xdist==3.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-json-logger==4.0.0
python-rapidjson==1.22
pytorch-lightning==2.5.5
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
qcs-api-client-common==0.11.8
qcs-sdk-python==0.21.20
quil==0.17.0
ray==2.49.2
referencing==0.36.2
regex==2025.10.23
reportlab==4.4.4
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.2.0
rpcq==3.11.0
rpds-py==0.27.1
rsa==4.9.1
ruamel.yaml==0.18.16
ruamel.yaml.clib==0.2.14
scipy==1.13.1
Send2Trash==2.0.0
sentry-sdk==2.43.0
setuptools-scm==9.2.2
sgmllib3k==1.0.0
shapely==2.0.7
shellingham==1.5.4
six==1.17.0
smart_open==7.4.1
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8.1
SQLAlchemy==2.0.44
sqlalchemy-spanner==1.17.1
sqlmodel==0.0.27
sqlparse==0.5.3
st-paywall==1.0.2
stack-data==0.6.3
starlette==0.48.0
stdlib-list==0.11.1
streamlit==1.50.0
stripe==13.0.1
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
tensorboardX==2.6.4
termcolor==3.1.0
terminado==0.18.1
tiktoken==0.12.0
tinycss2==1.4.0
tinyhtml5==2.0.0
tokenizers==0.22.2
toml==0.10.2
tomli==2.3.0
torch==2.2.2
torchmetrics==1.8.2
torchvision==0.17.2
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
typer==0.20.0
typer-slim==0.21.1
types-Deprecated==1.2.15.20250304
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
uri-template==1.3.0
uritemplate==4.2.0
urllib3==2.5.0
uuid_utils==0.14.0
uvicorn==0.38.0
uvloop==0.22.1
virtualenv==20.35.3
wandb==0.22.3
watchdog==6.0.0
watchfiles==1.1.1
wcwidth==0.2.14
weasyprint==66.0
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.9.0
websockets==15.0.1
wrapt==2.0.0
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
zopfli==0.2.3.post1
zstandard==0.25.0
